+++ internal/config/config.go
@@ -17,6 +17,7 @@ func GetConfig() error {
 	viper.SetDefault("Directory.Assets", "./web/assets")
 	viper.SetDefault("Symbolicator.Executable", "./minidump_stackwalk")
 	viper.SetDefault("Symbolicator.TrimModuleNames", true)
+	viper.SetDefault("Housekeeping.ReportRetentionTime", "2190h") // Around 3 months (duration only supports times in hours and down due to irregular length of days/months/years)
 
 	viper.SetConfigName("config")
 	viper.AddConfigPath("/etc/crashdragon/")
@@ -27,5 +28,5 @@ func GetConfig() error {
 		return err
 	}
 
-	return viper.SafeWriteConfig()
+	return viper.WriteConfig()
 }